+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Innotop for MySQL 8 written in Go
 - missing <K> option in help
 - remove panic in kill.go
 - display the error message when a thread_id is not available to be killed
+- simplify label prints in details
 
 ## 0.1.0 2021-04-06
 - Initial Release